tools/auto-preview/requirements.txt
robotpy-ntcore>=2026.0.0
Pillow>=10.0.0
